Export fixes.
Former-commit-id: 7b53eac378e1f8178e1dcdd8ba3bd65a93dda649

+++ app/services/sandbox/ExportStatic.scala
@@ -49,6 +49,8 @@ object ExportStatic {
   }
 
   private[this] def render(filename: String, content: Html) = {
-    Files.write(outPath.resolve(filename), content.toString().getBytes("UTF-8"))
+    if(Files.exists(outPath.resolve(filename))) {
+      Files.write(outPath.resolve(filename), content.toString().getBytes("UTF-8"))
+    }
   }
 }
